GUI-1/requirements.txt
flask
ultralytics
torch
pillow
joblib
pandas
scikit-learn
pytorch-tabnet

